Xenial: pick the right Lua
Now that I think about this, this is not ideal. It would be nice
if Debian-based distro's just had a lua.pc file for the latest
version of Lua. One can dream...
Instead I'll modify configure.ac to look for more different Lua
versions and this will all go away.