Handle .so files in new location
I'm working on a PR (https://github.com/collectd/collectd/pull/2085)
that is going to switch to a single non-recursive Makefile.

This changes the $(top_srcdir) and $(top_builddir), and as a result, the
location where .so files will be stored.

To test this in Jenkins, also look in this new location for .so files.